# fix and optimize varint encoding