I was looking into optimizing varint decoding because we have a use case where during CPU-intensive work, ~15% of the time is spent on this code path.

While writing the initial benchmarks to validate the change I noticed that encoding and decoding was broken for large integer values due to some mistakes in the way zig-zag encoding was being handled. This probably never triggered in practice because the varints are not used to encode values in the 2^56~2^64 range, but we're better off with a fix anyways ;)

The optimization yielded good results on the read path:
```
benchmark                old ns/op     new ns/op     delta
BenchmarkWriteVarInt     20.8          20.3          -2.40%
BenchmarkReadVarInt      93.3          38.0          -59.27%
```